requirements.txt
requests>=2.28.1,<3.0.0
python-dotenv==0.19.1